Make OpenAPI endpoint definitions authoritative
